+++ lib/src/commands/sdk_commands/sdk_command.dart
@@ -1,7 +1,9 @@
 import 'package:args/command_runner.dart';
+import 'package:flutter_compile/src/commands/sdk_commands/_sdk_global.dart';
 import 'package:flutter_compile/src/commands/sdk_commands/_sdk_install.dart';
 import 'package:flutter_compile/src/commands/sdk_commands/_sdk_list.dart';
 import 'package:flutter_compile/src/commands/sdk_commands/_sdk_remove.dart';
+import 'package:flutter_compile/src/commands/sdk_commands/_sdk_use.dart';
 import 'package:mason_logger/mason_logger.dart';
 
 /// {@template sdk_command}
@@ -10,6 +12,8 @@ import 'package:mason_logger/mason_logger.dart';
 /// `flutter_compile sdk install <version|channel>`
 /// `flutter_compile sdk list`
 /// `flutter_compile sdk remove <version>`
+/// `flutter_compile sdk global [version]`
+/// `flutter_compile sdk use [version]`
 ///
 /// A [Command] to manage multiple Flutter SDK versions.
 ///
@@ -20,6 +24,8 @@ class SdkCommand extends Command<int> {
     addSubcommand(SdkInstallSubCommand(_logger));
     addSubcommand(SdkListSubCommand(_logger));
     addSubcommand(SdkRemoveSubCommand(_logger));
+    addSubcommand(SdkGlobalSubCommand(_logger));
+    addSubcommand(SdkUseSubCommand(_logger));
   }
 
   @override